+++ bukkit/src/main/java/com/github/games647/changeskin/bukkit/listener/BungeeCordListener.java
@@ -6,6 +6,8 @@ import com.github.games647.changeskin.core.SkinData;
 import com.google.common.io.ByteArrayDataInput;
 import com.google.common.io.ByteStreams;
 
+import java.util.logging.Level;
+
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.messaging.PluginMessageListener;
@@ -24,15 +26,11 @@ public class BungeeCordListener implements PluginMessageListener {
             return;
         }
 
+        plugin.getLogger().log(Level.INFO, "Received instant update request from BungeeCord. "
+                + "This request should only be send if the command /setskin was invoked");
+
         ByteArrayDataInput dataInput = ByteStreams.newDataInput(message);
         String subchannel = dataInput.readUTF();
-
-//        UUID receiverUUID = UUID.fromString(dataInput.readUTF());
-//        Player receiver = Bukkit.getPlayer(receiverUUID);
-//        if (receiver == null) {
-//            plugin.getLogger().warning("BungeeCord requested Skin update, but receiver player isn't online");
-//            return;
-//        }
 
         String encodedData = dataInput.readUTF();
         if (encodedData.equalsIgnoreCase("null")) {
@@ -40,10 +38,18 @@ public class BungeeCordListener implements PluginMessageListener {
             return;
         }
 
+        Player receiver = player;
+        try {
+            String playerName = dataInput.readUTF();
+            receiver = Bukkit.getPlayerExact(playerName);
+            plugin.getLogger().log(Level.INFO, "Instant update for ", playerName);
+        } catch (Exception ex) {
+            plugin.getLogger().warning("You are using an outdated ChangeSkin spigot version");
+        }
 
         String signature = dataInput.readUTF();
         
         SkinData skinData = new SkinData(encodedData, signature);
-        Bukkit.getScheduler().runTask(plugin, new SkinUpdater(plugin, null, player, skinData));
+        Bukkit.getScheduler().runTask(plugin, new SkinUpdater(plugin, null, receiver, skinData));
     }
 }